+++ e2e/cypress/schemas/index.js
@@ -1,14 +1,21 @@
 import { combineSchemas } from '@cypress/schema-tools';
-import courseVersions from './course-schema';
+
 import articleVersions from './article-schema';
+import breadCrumbVersions from './breadcrumb-schema';
 import blogVersions from './blog-schema';
+import courseVersions from './course-schema';
+import localBusiness from './local-business-schema';
+import logoVersions from './logo-schema';
 import productVersions from './product-schema';
 import socialProfileVersions from './social-profile-schema';
 
 const schemas = combineSchemas(
-  courseVersions,
   articleVersions,
+  breadCrumbVersions,
   blogVersions,
+  courseVersions,
+  localBusiness,
+  logoVersions,
   productVersions,
   socialProfileVersions,
 );