+++ e2e/pages/jsonld.jsx
@@ -1,8 +1,11 @@
 import React from 'react';
 import {
   ArticleJsonLd,
+  BreadcrumbJsonLd,
   BlogJsonLd,
   CourseJsonLd,
+  LocalBusinessJsonLd,
+  LogoJsonLd,
   ProductJsonLd,
   SocialProfileJsonLd,
 } from '../../dist';
@@ -27,6 +30,31 @@ export default () => (
       description="This is a mighty good description of this article."
     />
 
+    <BreadcrumbJsonLd
+      itemListElements={[
+        {
+          position: 1,
+          name: 'Books',
+          item: 'https://example.com/books',
+        },
+        {
+          position: 2,
+          name: 'Authors',
+          item: 'https://example.com/books/authors',
+        },
+        {
+          position: 3,
+          name: 'Ann Leckie',
+          item: 'https://example.com/books/authors/annleckie',
+        },
+        {
+          position: 4,
+          name: 'Ancillary Justice',
+          item: 'https://example.com/books/authors/ancillaryjustice',
+        },
+      ]}
+    />
+
     <BlogJsonLd
       url="https://example.com/blog"
       title="Blog headline"
@@ -46,6 +74,36 @@ export default () => (
       providerName="Course Provider"
       providerUrl="https//www.example.com/provider"
       description="Course description goes right here"
+    />
+
+    <LocalBusinessJsonLd
+      type="Store"
+      id="http://davesdeptstore.example.com"
+      name="Dave's Department Store"
+      description="Dave's latest department store in San Jose, now open"
+      url="http://www.example.com/store-locator/sl/San-Jose-Westgate-Store/1427"
+      telephone="+14088717984"
+      address={{
+        streetAddress: '1600 Saratoga Ave',
+        addressLocality: 'San Jose',
+        addressRegion: 'CA',
+        postalCode: '95129',
+        addressCountry: 'US',
+      }}
+      geo={{
+        latitude: '37.293058',
+        longitude: '-121.988331',
+      }}
+      images={[
+        'https://example.com/photos/1x1/photo.jpg',
+        'https://example.com/photos/4x3/photo.jpg',
+        'https://example.com/photos/16x9/photo.jpg',
+      ]}
+    />
+
+    <LogoJsonLd
+      logo="http://www.your-site.com/images/logo.jpg"
+      url="http://www.your-site.com"
     />
 
     <ProductJsonLd