+++ src/index.js
@@ -1,7 +1,10 @@
 import DefaultSeo from './meta/defaultSEO';
 import ArticleJsonLd from './jsonld/article';
+import BreadcrumbJsonLd from './jsonld/breadcrumb';
 import BlogJsonLd from './jsonld/blog';
 import CourseJsonLd from './jsonld/course';
+import LocalBusinessJsonLd from './jsonld/localBusiness';
+import LogoJsonLd from './jsonld/logo';
 import ProductJsonLd from './jsonld/product';
 import SocialProfileJsonLd from './jsonld/socialProfile';
 
@@ -9,7 +12,10 @@ export default DefaultSeo;
 export {
   ArticleJsonLd,
   BlogJsonLd,
+  BreadcrumbJsonLd,
   CourseJsonLd,
+  LocalBusinessJsonLd,
+  LogoJsonLd,
   ProductJsonLd,
   SocialProfileJsonLd,
 };