+++ src/meta/__tests__/buildTags.spec.jsx
@@ -1,5 +1,4 @@
 import { getByText, render } from 'react-testing-library';
-import React from 'react';
 
 import buildTags from '../buildTags';